Move xml_util.py into test_formatters
