+++ rdm/translate.py
@@ -1,5 +1,5 @@
 from rdm.util import write_yaml
-from rdm.xml_util import flattened_gtest_results, xml_load, flattened_qttest_results, auto_translator
+from rdm.test_formatters.xml_util import flattened_gtest_results, xml_load, flattened_qttest_results, auto_translator
 
 XML_TRANSLATORS = {
     'auto': auto_translator,
