+++ tests/test_xml_util.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from rdm.xml_util import xml_load, flattened_gtest_results, flattened_qttest_results, auto_translator
+from rdm.test_formatters.xml_util import xml_load, flattened_gtest_results, flattened_qttest_results, auto_translator
 
 
 def _full_path_of_test_file(name):
